feat: added edit crop image

+++ src/components/layout/right-side-bar/theme-tab/colorScheme/index.tsx
@@ -4,6 +4,17 @@ import { PlusIcon, Trash } from 'lucide-react';
 import { HexColorPicker } from 'react-colorful';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { Input } from '@/components/ui/input';
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from '@/components/ui/alert-dialog';
 
 export default function ColorScheme() {
   const { colorScheme, addNewColor } = useColorScheme();
@@ -31,7 +42,10 @@ function ColorName({ colorName, colorHex, index }: { colorName: string; colorHex
     <div>
       <div
         className="visio-cms-flex visio-cms-group visio-cms-bg-dark-900 visio-cms-rounded-md visio-cms-p-2 visio-cms-mb-3 visio-cms-items-center visio-cms-justify-between hover:visio-cms-bg-dark-700 visio-cms-cursor-pointer"
-        onClick={() => setIsEditing(true)}
+        onClick={(e) => {
+          e.stopPropagation();
+          setIsEditing(true);
+        }}
       >
         <div className="visio-cms-flex visio-cms-gap-2 visio-cms-items-center">
           <Popover>
@@ -74,16 +88,36 @@ function ColorName({ colorName, colorHex, index }: { colorName: string; colorHex
           )}
         </div>
 
-        <Button
-          variant={'ghost'}
-          className="visio-cms-invisible group-hover:visio-cms-visible"
-          onClick={(e) => {
-            e.stopPropagation();
-            deleteColor(index);
-          }}
-        >
-          <Trash size={15} />
-        </Button>
+        <AlertDialog>
+          <AlertDialogTrigger
+            asChild
+            onClick={(e) => {
+              e.stopPropagation();
+            }}
+          >
+            <Button variant={'ghost'} className="visio-cms-invisible group-hover:visio-cms-visible">
+              <Trash size={15} />
+            </Button>
+          </AlertDialogTrigger>
+          <AlertDialogContent>
+            <AlertDialogHeader>
+              <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+              <AlertDialogDescription>
+                This action cannot be undone. This will permanently delete all instances of this color in your project.
+              </AlertDialogDescription>
+            </AlertDialogHeader>
+            <AlertDialogFooter>
+              <AlertDialogCancel className="visio-cms-text-white">Cancel</AlertDialogCancel>
+              <AlertDialogAction
+                onClick={() => {
+                  deleteColor(index);
+                }}
+              >
+                Continue
+              </AlertDialogAction>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialog>
       </div>
     </div>
   );
